Add another decay feature vector based on a Gaussian distribution
This feature vector works the same as the other decay feature vector
with the alteration that it uses a new decay function. This was picked
because after doing PCA on the binary feature vector it seemed as though
there was a slight Gaussian trend in the influence of each layer of the
name.