Add segmenting and other per property visualizations to pca_plot
These allow the plotting of the progression of each of the properties
over time rather than just looking at the whole blob of properties.

With segmentting, this is a much higher level view where each property
is split in half and "high" values get a 1 for their respective color,
and "low" values get a 0 for their respective color.